[IOTDB-3143] Potential bug fix for library-udf (#5857)

+++ library-udf/src/main/java/org/apache/iotdb/library/dprofile/UDTFQLB.java
@@ -89,7 +89,7 @@ public class UDTFQLB implements UDTF {
         }
       }
       correlation = correlation / n;
-      collector.putDouble(n + shift, correlation);
+      collector.putDouble((long) n + shift, correlation);
       qlb += correlation * correlation / (n - shift) * n * (n + 2);
       ChiSquaredDistribution qlbdist = new ChiSquaredDistribution(shift);
       double qlbprob = 1.0 - qlbdist.cumulativeProbability(qlb);
